Don't migrate if the newer path already exists

+++ taskcoach/taskcoachlib/config/settings.py
@@ -336,16 +336,17 @@ class Settings(object, CachingConfigParser):
         else:
             path = os.path.join(path, *args)
 
-        if not os.path.exists(path):
+        exists = os.path.exists(path)
+        if not exists:
             os.makedirs(path)
-        return path
+        return path, exists
 
     def pathToTemplatesDir(self):
         try:
             return self.pathToDataDir('templates')
         except:
             pass # Fallback on old path
-        return self.pathToTemplatesDir_deprecated()
+        return self.pathToTemplatesDir_deprecated(), True
 
     def pathToConfigDir_deprecated(self, environ):
         try:
@@ -387,7 +388,9 @@ class Settings(object, CachingConfigParser):
     def migrateConfigurationFiles(self):
         # Templates. Extra care for Windows shortcut.
         oldPath = os.path.join(self.pathToConfigDir_deprecated(environ=os.environ), 'taskcoach-templates')
-        newPath = self.pathToTemplatesDir()
+        newPath, exists = self.pathToTemplatesDir()
+        if exists:
+            return
         if oldPath != newPath:
             if operating_system.isWindows() and os.path.exists(oldPath + '.lnk'):
                 shutil.move(oldPath + '.lnk', newPath + '.lnk')
